requirements.txt
Package            Version
------------------ ---------
blinker            1.4
brotlipy           0.7.0
certifi            2021.10.8
cffi               1.14.6
chardet            4.0.0
charset-normalizer 2.0.0
click              8.0.3
colorama           0.4.4
cryptography       35.0.0
cycler             0.11.0
idna               3.1
joblib             1.1.0
kiwisolver         1.3.2
matplotlib         3.4.3
nltk               3.6.5
numpy              1.21.3
oauthlib           3.1.1
olefile            0.46
pandas             1.3.4
Pillow             8.3.2
pip                21.3.1
pycparser          2.20
PyJWT              2.3.0
pyOpenSSL          21.0.0
pyparsing          3.0.5
PyQt5              5.12.3
PyQt5_sip          4.19.18
PyQtChart          5.12
PyQtWebEngine      5.12.1
PySocks            1.7.1
python-dateutil    2.8.2
python-dotenv      0.19.1
pytz               2021.3
regex              2021.11.2
requests           2.26.0
requests-oauthlib  1.3.0
setuptools         58.4.0
six                1.16.0
tornado            6.1
tqdm               4.62.3
tweepy             4.2.0
urllib3            1.26.7
wheel              0.37.0
win-inet-pton      1.1.0
